+++ src/test/java/com/senzing/api/services/HowRelatedServicesTest.java
@@ -11,6 +11,8 @@ import com.senzing.util.JsonUtilities;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.TestInstance;
+import org.junit.jupiter.api.parallel.Execution;
+import org.junit.jupiter.api.parallel.ExecutionMode;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
@@ -36,6 +38,7 @@ import static org.junit.jupiter.params.provider.Arguments.arguments;
 import static com.senzing.io.IOUtilities.*;
 
 @TestInstance(Lifecycle.PER_CLASS)
+@Execution(ExecutionMode.SAME_THREAD)
 public class HowRelatedServicesTest extends AbstractServiceTest {
   private static final long RANDOM_SEED = 1234567890L;