+++ src/test/java/com/senzing/api/services/ReadOnlyEntityDataWriteServicesTest.java
@@ -6,6 +6,8 @@ import com.senzing.api.server.SzApiServerOptions;
 import com.senzing.util.JsonUtilities;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.TestInstance;
+import org.junit.jupiter.api.parallel.Execution;
+import org.junit.jupiter.api.parallel.ExecutionMode;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
 import org.springframework.web.client.HttpStatusCodeException;
@@ -25,6 +27,7 @@ import static org.junit.jupiter.api.TestInstance.Lifecycle;
 import static com.senzing.api.services.ResponseValidators.*;
 
 @TestInstance(Lifecycle.PER_CLASS)
+@Execution(ExecutionMode.SAME_THREAD)
 public class ReadOnlyEntityDataWriteServicesTest
     extends EntityDataWriteServicesTest
 {